gnhf 18: Extracted a shared PreferenceExample::from_row mapper in src/taste/compiler.rs and routed the two duplicated row-construction blocks in taste/export.rs and taste/show.rs through it by standardizing their SQL SELECT column order, with no functional change.
