+++ src/taste/show.rs
@@ -5,7 +5,7 @@ use crate::JottraceError;
 use crate::storage::{query_collect, sqlite_error};
 use crate::{Result, data_dir_from_env, open_locked_database};
 
-use super::compiler::{EvidenceKind, PreferenceExample, PreferenceOutcome};
+use super::compiler::PreferenceExample;
 use super::timeline::{FileTimelineRow, TimelineSourceKind, normalize_file_path};
 
 const CLAUDE_SOURCE: &str = "claude_cli";
@@ -153,23 +153,25 @@ fn query_preference_example(
         Some(source_session_id) => query_collect(
             db_path,
             conn,
-            "SELECT source_session_id, generation, proposal_event_seq, file_path, tool_name,
-                    proposal_content, context, outcome, confidence, evidence_kind, extractor_version
+            "SELECT source, source_session_id, generation, proposal_event_seq, tool_use_id,
+                    file_path, tool_name, proposal_content, context, outcome, confidence,
+                    evidence_kind, extractor_version
              FROM preference_examples
              WHERE source = ?1 AND source_session_id = ?2 AND tool_use_id = ?3",
             params![source, source_session_id, tool_use_id],
-            |row| map_preference_row(source, tool_use_id, row),
+            PreferenceExample::from_row,
         )?,
         None => query_collect(
             db_path,
             conn,
-            "SELECT source_session_id, generation, proposal_event_seq, file_path, tool_name,
-                    proposal_content, context, outcome, confidence, evidence_kind, extractor_version
+            "SELECT source, source_session_id, generation, proposal_event_seq, tool_use_id,
+                    file_path, tool_name, proposal_content, context, outcome, confidence,
+                    evidence_kind, extractor_version
              FROM preference_examples
              WHERE source = ?1 AND tool_use_id = ?2
              ORDER BY source_session_id ASC",
             params![source, tool_use_id],
-            |row| map_preference_row(source, tool_use_id, row),
+            PreferenceExample::from_row,
         )?,
     };
 
@@ -183,34 +185,6 @@ fn query_preference_example(
             session_count: rows.len(),
         }),
     }
-}
-
-fn map_preference_row(
-    source: &str,
-    tool_use_id: &str,
-    row: &rusqlite::Row<'_>,
-) -> rusqlite::Result<PreferenceExample> {
-    let source_session_id: String = row.get(0)?;
-    let generation: i64 = row.get(1)?;
-    let proposal_event_seq: i64 = row.get(2)?;
-    let outcome: String = row.get(7)?;
-    let evidence_kind: String = row.get(9)?;
-    Ok(PreferenceExample {
-        source: source.to_string(),
-        source_session_id,
-        generation: usize::try_from(generation).expect("generation fits in usize"),
-        proposal_event_seq: usize::try_from(proposal_event_seq)
-            .expect("proposal_event_seq fits in usize"),
-        tool_use_id: tool_use_id.to_string(),
-        file_path: row.get(3)?,
-        tool_name: row.get(4)?,
-        proposal_content: row.get(5)?,
-        context: row.get(6)?,
-        outcome: PreferenceOutcome::from_db_str(&outcome).expect("valid outcome"),
-        confidence: row.get(8)?,
-        evidence_kind: EvidenceKind::from_db_str(&evidence_kind).expect("valid evidence_kind"),
-        extractor_version: row.get(10)?,
-    })
 }
 
 fn query_timeline_rows(
